+++ tests/test_binaries.py
@@ -28,6 +28,7 @@ class TestBinaryPackage(unittest.TestCase):
         )
         self.assertTrue([('Running', {'EIP': 4196624})], list(f.threads()))
         self.assertIsNone(f.getInterpreter())
+        f.elf.stream.close()
 
     def test_decree(self):
         filename = os.path.join(os.path.dirname(__file__), 'binaries', 'cadet_decree_x86')
@@ -37,6 +38,7 @@ class TestBinaryPackage(unittest.TestCase):
             list(f.maps())
         )
         self.assertTrue([('Running', {'EIP': 134513708})], list(f.threads()))
+        f.elf.stream.close()
 
 
 class IntegrationTest(unittest.TestCase):
@@ -139,7 +141,8 @@ class IntegrationTest(unittest.TestCase):
         filename = filename[len(os.getcwd())+1:]
         workspace = os.path.join(self.test_dir, 'workspace')
         assertions = os.path.join(self.test_dir, 'assertions.txt')
-        open(assertions,'w').write('0x0000000000401003 ZF == 1')
+        with open(assertions, 'w') as output:
+            output.write('0x0000000000401003 ZF == 1')
         with open(os.path.join(os.pardir, self.test_dir, 'output.log'), "w") as output:
             subprocess.check_call([PYTHON_BIN, '-m', 'manticore',
                                    '--workspace', workspace,